[sys_mon] Sys_mon and rtos_mon merged into one module, added support for CPU load on SIM targets
