+++ conf/airframes/ENAC/fixed-wing/jp.xml
@@ -69,7 +69,7 @@
     <module name="current_sensor">
       <configure name="ADC_CURRENT_SENSOR" value="ADC_2"/>
     </module>
-    <!--module name="rtos_mon"/-->
+    <!--module name="sys_mon"/-->
   </firmware>
 
   <!-- commands section -->
