+++ sw/airborne/arch/chibios/modules/core/rtos_mon_arch.c
@@ -24,11 +24,9 @@
  * ChibiOS implementation
  */
 
-#include "modules/core/rtos_mon.h"
+#include "modules/core/sys_mon_rtos.h"
 #include "modules/core/rtos_mon_arch.h"
-#include "subsystems/datalink/downlink.h"
 #include <ch.h>
-#include <string.h>
 
 #if !CH_DBG_THREADS_PROFILING
 #error CH_DBG_THREADS_PROFILING should be defined to TRUE to use this monitoring tool
@@ -91,11 +89,9 @@ void rtos_mon_periodic_arch(void)
 
   // assume we call the counter once a second
   // so the difference in seconds is always one
-  // NOTE: not perfectly precise
-  // FIXME: add finer resolution than seconds?
+  // NOTE: not perfectly precise, +-5% on average so take it into consideration
   rtos_mon.cpu_load = (1 - (float)(idle_counter - last_idle_counter) / CH_CFG_ST_FREQUENCY) * 100;
   last_idle_counter = idle_counter;
-
 }
 
 static uint16_t get_stack_free(const thread_t *tp)
