+++ sw/airborne/arch/sim/modules/core/rtos_mon_arch.h
@@ -18,10 +18,10 @@
  * <http://www.gnu.org/licenses/>.
  */
 /**
- * @file "arch/chibios/modules/core/rtos_mon_arch.h"
+ * @file "arch/sim/modules/core/rtos_mon_arch.h"
  * @author Gautier Hattenberger
- * RTOS monitoring tool
- * ChibiOS implementation
+ * System monitoring tool
+ * SIM implementation
  */
 
 #ifndef RTOS_MON_ARCH_H
