+++ sw/airborne/modules/core/sys_mon_bare_metal.h
@@ -26,28 +26,23 @@
  * return cpu load, average exec time, ...
  */
 
-#ifndef SYS_MON_H
-#define SYS_MON_H
+#ifndef SYS_MON_BARE_METAL_H
+#define SYS_MON_BARE_METAL_H
 
-#include "std.h"
+#include "core/sys_mon.h"
+
+struct SysMon {
+  uint8_t  cpu_load;
+  uint16_t periodic_time;      ///< in usec
+  uint16_t periodic_time_min;  ///< in usec
+  uint16_t periodic_time_max;  ///< in usec
+  uint16_t periodic_cycle;     ///< in usec
+  uint16_t periodic_cycle_min; ///< in usec
+  uint16_t periodic_cycle_max; ///< in usec
+  uint16_t event_number;
+  float cpu_time; // in secs since startup
+};
 
 extern struct SysMon sys_mon;
 
-/** Init system monitoring
- */
-void init_sysmon(void);
-
-/** Report system status
- */
-void periodic_report_sysmon(void);
-
-/** Analyse periodic calls
- *  Should be run at the highest frequency
- */
-void periodic_sysmon(void);
-
-/** Analyse event calls
- */
-void event_sysmon(void);
-
-#endif
+#endif /* SYS_MON_BARE_METAL_H */
